+++ idea/testData/intentions/convertEnumToSealedClass/membersOnly.kt
@@ -1,0 +1,9 @@
+// WITH_RUNTIME
+
+enum class <caret>MyEnum(val s: String = "") {
+    ;
+
+    fun foo() {
+
+    }
+}